requirements.txt
PyPDF2>=3.0.0
pdfplumber>=0.9.0
pdf2image>=1.16.0
pytesseract>=0.3.10
Pillow>=9.0.0